environment in place
I am done working on this for today.
I got all the environment in (unless I want to add a bit more)
add the colliders and import all the other elements from testing and we should be good. You can move around the platforms if you want, they are just sprites.